@@ -8,22 +8,26 @@
   <packaging>${packaging}</packaging>
 
   <parent>
-    <groupId>io.micronaut</groupId>
+    <groupId>io.micronaut.platform</groupId>
     <artifactId>micronaut-parent</artifactId>
-    <version>3.10.4</version>
+    <version>4.6.3</version>
   </parent>
 
   <properties>
     <packaging>jar</packaging>
-    <jdk.version>17</jdk.version>
-    <release.version>17</release.version>
+    <jdk.version>21</jdk.version>
+    <release.version>21</release.version>
     <micronaut.runtime>netty</micronaut.runtime>
     <exec.mainClass>org.signal.cdsi.Application</exec.mainClass>
 
-    <testcontainers.version>1.19.7</testcontainers.version>
-    <aws.sdk.version>2.25.34</aws.sdk.version>
-    <commons.codec.version>1.16.1</commons.codec.version>
-    <guava.version>33.1.0-jre</guava.version>
+    <testcontainers.version>1.20.4</testcontainers.version>
+    <aws.sdk.version>2.29.45</aws.sdk.version>
+    <commons.codec.version>1.17.1</commons.codec.version>
+    <guava.version>33.4.0-jre</guava.version>
+    <make.docker.prefix>docker_</make.docker.prefix>
+    <httpcomponents.httpcore.version>4.4.16</httpcomponents.httpcore.version>
+    <hdrhistogram.version>2.2.2</hdrhistogram.version>
+    <jetty.version>12.0.16</jetty.version>
   </properties>
 
   <repositories>
