@@ -7,8 +7,8 @@ package org.signal.cdsi.account.aws;
 
 import io.micronaut.context.annotation.ConfigurationProperties;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Positive;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Positive;
 
 @ConfigurationProperties("accountTable")
 class AccountTableConfiguration {
