@@ -7,7 +7,7 @@ package org.signal.cdsi.limits.cosmos;
 
 import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.context.annotation.Context;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @ConfigurationProperties("cosmos")
 @Context
